requirements.txt
scikit-learn
scipy
numpy
joblib
GitPython
